Use sdb-admin to set global variables
We were simply setting global variables in the MA using memsqlctl. This of course means we are not setting variables for leaves.
This isn't a problem for most recent variables, which are sync vars, but can be a problem for non-sync ones.

In my case, I was running the singlestore-dev image in a pretty big machine, which made the default maximum_memory too big for the free license. Setting the maximum_memory at the MA level wasn't enough.