Add SingleStore Kai support (#98)
This PR adds SingleStore Kai to the dev image. Including the binaries
adds ~30 MiB to the footprint. It is optional and defaults to OFF, it
must be enabled with ENABLE_KAI=1.

This PR replaces the "extension" mechanism which was not documented but
used to run Kai in earlier iterations of the dev image.