Add singlestore / mongosh clients that authenticate using env var (#104)
The current sinlgestore client in the container requires that you
specify a password even though the root password is already available as
an env var. I've added a new client interface (singlestore-auth) and new
mongosh client (mongosh-auth) that use the existing password so that you
do not get prompted for it when using these interfaces.