fix: enable auto-tag to fetch with full history (#123)
The change fixes the failure in the auto-tag job, where, in the recent
runs, I found that the tag wasn't listed properly. The reason was the
default checkout action did not include fetching with tags. Add an
option to fetch with depth `0`, so that it will include full history
with tags.